Migrate to_subgait_msg to subgait class
